Scripts/Raspberry/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
numpy==2.2.2
paho-mqtt==1.6.1
pandas==2.2.3
pyftdi==0.55.4
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.1
pyusb==1.2.1
RPi.GPIO==0.7.1
scd30-i2c==0.0.6
six==1.17.0
smbus2==0.3.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
tzdata==2025.1
